requirements.txt
asgiref==3.11.1
click==8.3.1
colorama==0.4.6
croniter==6.2.2
Django==6.0.3
django-cors-headers==4.9.0
django-rq==4.0.1
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==40.11.0
flake8==7.3.0
mccabe==0.7.0
pillow==12.1.1
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==7.3.0
rq==2.7.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
